Pull in more MrOne test branch changes
https://jihulab.com/mr-one/openpilot/-/commit/3fae24d538fb94891be1a6ffa56c68334160dc88
but if we're getting rid of `NetworkLocation.fwdCamera` for `NetworkLocation.gateway`
in one place I also updated in the other for the Volkswagen car.
This change may be useless and I don't have a Volkswagen so it's hard to say if it's a
good change or not

https://jihulab.com/mr-one/openpilot/-/commit/b194f775523686c94fc095fb5de20b5b72cb357e
Enables back some of the commented out service like `ubloxd` and `locationd`
`micd` and `soundd` are still commented out and `ubloxd` was uncommented then commeted again
No idea why, but in this commit I've enabled `ubloxd` because I want to see if any
issues arise.